refactor(core): move change password form in settings to own page
