+++ .changeset/soft-brooms-know.md
@@ -1,0 +1,5 @@
+---
+"@bigcommerce/catalyst-core": minor
+---
+
+Add /account/settings/change-password route for change password form.
