+++ core/app/[locale]/(default)/account/(tabs)/_components/account-status-provider.tsx
@@ -3,7 +3,7 @@
 import { usePathname } from 'next/navigation';
 import { createContext, ReactNode, useContext, useEffect, useState } from 'react';
 
-import { State as AccountState } from '../settings/_actions/submit-customer-change-password-form';
+import { State as AccountState } from '../settings/change-password/_actions/change-password';
 
 const defaultState: AccountState = { status: 'idle', message: '' };
