+++ core/app/[locale]/(default)/account/(tabs)/settings/page.tsx
@@ -2,19 +2,10 @@ import { notFound } from 'next/navigation';
 
 import { TabHeading } from '../_components/tab-heading';
 
-import { SettingsContent } from './_components/settings-content';
+import { UpdateSettingsForm } from './_components/update-settings-form';
 import { getCustomerSettingsQuery } from './page-data';
 
-interface Props {
-  searchParams: {
-    [key: string]: string | string[] | undefined;
-    action?: 'change_password';
-    before?: string;
-    after?: string;
-  };
-}
-
-export default async function Settings({ searchParams }: Props) {
+export default async function Settings() {
   const customerSettings = await getCustomerSettingsQuery({
     address: { filters: { entityIds: [4, 5, 6, 7] } },
   });
@@ -26,7 +17,9 @@ export default async function Settings({ searchParams }: Props) {
   return (
     <>
       <TabHeading heading="settings" />
-      <SettingsContent action={searchParams.action} customerSettings={customerSettings} />
+      <div className="mx-auto lg:w-2/3">
+        <UpdateSettingsForm {...customerSettings} />
+      </div>
     </>
   );
 }
